Handle missing padding after last attribute
The requirement is for each attribute to *start* on word boundary, but
not necessarily end, it seems to be legal for the last attribute in a
message (and buffer) not to be followed by a padding. Apparently at
least the NFQUEUE takes advantage and sends such messages.

This relaxes the strictness about that and silently just assumes the end
of buffer was reached.