Fix size of ifa_flags field in Ifaddrmsg
Ifaddrmsg currently uses IfaFFlags for its ifa_flags field, but the
structure on the wire only has one-byte flags, not four-byte flags.
(Four-byte flags get sent as an attr.)

Fix the serialization and deserialization of Ifaddrmsg, Add From and
TryFrom impls to convert IfaFFlags to and from libc::c_uchar.