Bug fix for race condition and unpredictable presence of DONE packet.
This commit addresses two separate issues:

1. There was previously a short window where a message was sent to the
netlink socket, but the sequence number had not yet been registered
for routing back to the context that it was sent from. This caused
messages to periodically disappear unexpectedly.
2. There was some inconsistency in how Nlmsg::Done packets were handled.
This commit standardizes on always returning them from the iteratora.
While this may not be preferable from a user standpoint, it will
significantly simplify the code, resulting in less room for error in the
future.