fix: log naked structs at trace!() level
The problem with logging at debug level is that we have
neli as part of a much bigger binary, and right now neli
dumps very big unformatted structs in syslog.

We believe that neli should instead log these items at a
lower level, such as trace!(). Debug-level log messages
are still relevant for end users of the system, but trace
messages are only for developers that know their way
around on the system and thus are expected to know how
to filter noise out (e.g. via environment variables to
the logger).