Directly implement FromBytesWithInput and ToBytes for Buffer
Currently, Buffer will forward the implementations of FromBytesWithInput
and ToBytes to the generic implementation of Vec<_>. Is implementation
is quite inefficient when the contained element is small (like u8) and
there might be many elements (like a Buffer).

I don't see any way to specialise the Vec<u8> case, so I've implemented
FromBytesWithInput and ToBytes directly for Buffer.

A very simple benchmark with 1500 bytes show these results:
 - From Buffer: 27.94ns/call
 - From Vec<u8>: 862.63ns/call
 - To Buffer: 28.34ns/call
 - To Vec<u8>: 1288.43ns/call

This shows a massive improvement when Buffer is used.

Additionally, I've added Into<Vec<u8>> to Buffer to easily retrieve the
stored Vec.

Signed-off-by: Christian Speich <c.speich@avm.de>