Handle empty input in String's FromBytesWithInput implementation
This fixes "attempt to subtract with overflow" error when the input
buffer is empty. In this case

  buffer.position() + input - 1

is 0 + 0 - 1 leading to integer overflow.

In release builds integer overflows are not checked which could lead
into out of bounds read of the buffer.

While here validate UTF-8 string before allocating memory for it.